Move caption parsing out into another file
