+++ lib/mp4/probe.js
@@ -9,7 +9,8 @@
 'use strict';
 
 var toUnsigned = require('../utils/bin').toUnsigned;
-var findBox, parseType, timescale, startTime;
+var parseSei = require('../m2ts/captions-parser').parseSei;
+var findBox, parseType, timescale, startTime, findSeiNals, captionNals;
 
 // Find the data for a box specified by its path
 findBox = function(data, path) {
@@ -182,8 +183,52 @@ startTime = function(timescale, fragment) {
   return isFinite(result) ? result : 0;
 };
 
+findSeiNals = function(avcStream) {
+  var
+    avcView = new DataView(avcStream.buffer, avcStream.byteOffset, avcStream.byteLength),
+    result = [],
+    seiNal,
+    i,
+    length;
+  for (i = 0; i + 4 < avcStream.length; i += length) {
+    length = avcView.getUint32(i);
+    i += 4;
+
+    // bail if this doesn't appear to be an H264 stream
+    if (length <= 0) {
+      continue;
+    }
+
+    switch (avcStream[i] & 0x1F) {
+    case 0x06:
+      seiNal = {
+        type: 'sei_rbsp',
+        size: length,
+        nal: parseSei(avcStream.subarray(i + 1, i + 1 + length))
+      };
+      result.push(seiNal);
+      break;
+    default:
+      break;
+    }
+  }
+  return result;
+};
+
+captionNals = function(fragment) {
+  var captionNals = [];
+  var mdat = findBox(fragment, ['mdat']);
+
+  for (var i = 0; i < mdat.length; i++) {
+    captionNals = captionNals.concat(findSeiNals(mdat[i]));
+  }
+
+  return captionNals;
+}
+
 module.exports = {
   parseType: parseType,
   timescale: timescale,
-  startTime: startTime
+  startTime: startTime,
+  captionNals: captionNals
 };
