+++ lib/tools/mp4-inspector.js
@@ -14,7 +14,7 @@ var
   textifyMp4,
 
   parseType = require('../mp4/probe').parseType,
-  parseSei = require('../m2ts/caption-stream').parseSei,
+  parseSei = require('../m2ts/captions-parser').parseSei,
   parseMp4Date = function(seconds) {
     return new Date(seconds * 1000 - 2082844800000);
   },
